Main Loop Running (In the Console)
I got the main loop to start running in the terminal only. A few things needed to be fixed, so now the only thing that players can do in the loop is to move, get taxed, and pass Go. Pygame still is not rendering, so that will need to be the next step.